update files
Updated UI - HTML, CSS, and got database working. 
modified create, delete, details, edit, and main index for Quotes UI
modeified layout, home index file, desk.cs, and seeddata.cs to run with new megadesk parameters
updated site.css and bootstrap.css

+++ RazorPagesMegaDesk/RazorPagesMovie/Migrations/20181031000601_Initial.cs
@@ -18,7 +18,7 @@ namespace MegaDesk.Migrations
                     width = table.Column<double>(nullable: true),
                     depth = table.Column<double>(nullable: true),
                     numDrawers = table.Column<int>(nullable: true),
-                    price = table.Column<double>(nullable: true),
+                    price = table.Column<decimal>(nullable: true),
                     surfaceMat = table.Column<string>(nullable: true),
                     rushOrd = table.Column<string>(nullable: true),
                     DateAdded = table.Column<DateTime>(nullable: false)
